Revert "Refactor app architecture, clean up code, and minimize traffic"
