+++ .gitignore
@@ -1,18 +1,15 @@
+node_modules/
+.env
 answers_photos.csv
 answers.csv
 questions.csv
-
-
-node_modules/
-.env
-
-
+.DS_Store
 helpful.sql
 newrelic.js
-*.log
-*.pem
+newrelic_agent.log
+dbQuery.txt
+httpRequest.txt
+ec2-qa-postgres.pem
+catborg.pem
+deployment.md
 .*
-notes.md
-
-
-.DS_Store
