--- db/index.js
@@ -1,18 +1,0 @@
-const { Pool } = require('pg');
-require('dotenv').config();
-
-// Database connection
-const pool = new Pool({
-  user: process.env.DB_USER,
-  host: process.env.DB_HOST,
-  database: process.env.DB_DATABASE,
-  password: process.env.DB_PASSWORD,
-  port: process.env.DB_PORT,
-  max: 90,
-  idleTimeoutMillis: 2500,
-  connectionTimeoutMillis: 0,
-})
-
-module.exports = {
-  pool
-}
