+++ gateway.js
@@ -1,13 +1,14 @@
 const express = require('express');
-const router = require('./routes/index');
-const { pool } = require('./db/index');
-const fs = require('fs');
-
+const routes = require('./routes/index');
 const app = express();
 const PORT = 3000;
+const fs = require('fs');
+require('dotenv').config();
 
-// For loader io testing
+
+// For testing
 app.get(`/${process.env.LOADER_FILE}`, async (req, res) => {
+
   try {
     const data = fs.readFileSync(`.${process.env.LOADER_FILE}`, 'utf8')
     res.end(data);
@@ -15,18 +16,14 @@ app.get(`/${process.env.LOADER_FILE}`, async (req, res) => {
     console.error(err);
     res.end();
   }
-})
 
+})
 
 
 // Middleware
 app.use(express.json());
-app.use('/', router);
+app.use('/', routes);
 
 app.listen(PORT, () => {
-  pool.connect(() => {
-    console.log(`Database is connected on ${PORT}`);
-  });
-
   console.log(`Gateway listening on ${PORT}`)
 })
